# docs/swagger.yaml
basePath: /api/v1
definitions:
  controller._ResponseCommunityDetail:
    properties:
      code:
        $ref: '#/definitions/controller.ResCode'
      data:
        $ref: '#/definitions/models.CommunityDetail'
      message:
        type: string
    type: object
  controller._ResponseCommunityList:
    properties:
      code:
        $ref: '#/definitions/controller.ResCode'
      data:
        items:
          $ref: '#/definitions/models.Community'
        type: array
      message:
        type: string
    type: object
  controller._ResponsePostDetail:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/controller.ResCode'
        description: 业务响应状态码
      data:
        allOf:
        - $ref: '#/definitions/models.ApiPostDetail'
        description: 数据
      message:
        description: 提示信息
        type: string
    type: object
  controller._ResponsePostList:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/controller.ResCode'
        description: 业务响应状态码
      data:
        description: 数据
        items:
          $ref: '#/definitions/models.ApiPostDetail'
        type: array
      message:
        description: 提示信息
        type: string
    type: object
  controller.ResCode:
    enum:
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    type: integer
    x-enum-varnames:
    - CodeSuccess
    - CodeInvalidParam
    - CodeUserExist
    - CodeUserNotExist
    - CodeInvalidPassword
    - CodeServerBusy
    - CodeNeedLogin
    - CodeInvalidToken
  models.ApiPostDetail:
    properties:
      author_id:
        description: TODO:在请求中没有问题，但是会出现 响应类型不符的问题：在于json中的“string”
        type: integer
      author_name:
        description: 发帖作者的名称
        type: string
      community_id:
        type: integer
      content:
        type: string
      create_time:
        description: 这里如果要使用 time 类型的时间，注意连接库的时候，需要设置参数 parseTime=true
        type: string
      id:
        type: integer
      introduction:
        description: omitempty 是说如果这个字段为空，那么就不返回这个字段
        type: string
      name:
        type: string
      status:
        type: integer
      title:
        type: string
      vote_num:
        type: integer
    required:
    - community_id
    - content
    - title
    type: object
  models.Community:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.CommunityDetail:
    properties:
      create_time:
        description: 这里如果要使用 time 类型的时间，注意连接库的时候，需要设置参数 parseTime=true
        type: string
      id:
        type: integer
      introduction:
        description: omitempty 是说如果这个字段为空，那么就不返回这个字段
        type: string
      name:
        type: string
    type: object
host: 127.0.0.1:8080
info:
  contact:
    email: support@swagger.io
    name: lion
    url: http://www.swagger.io/support
  description: 这里写描述信息
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 这里写标题
  version: "1.0"
paths:
  /community:
    get:
      consumes:
      - application/json
      description: 查询社区列表
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseCommunityList'
      security:
      - ApiKeyAuth: []
      summary: 查询社区列表
      tags:
      - 社区相关接口
  /community/{id}:
    get:
      consumes:
      - application/json
      description: 查询社区详情列表
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      - description: 社区ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseCommunityDetail'
      security:
      - ApiKeyAuth: []
      summary: 查询社区详情列表
      tags:
      - 社区相关接口
  /post/{id}:
    get:
      consumes:
      - application/json
      description: 单个帖子详情接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      - description: 帖子id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponsePostDetail'
      security:
      - ApiKeyAuth: []
      summary: 单个帖子详情接口
      tags:
      - 帖子相关接口
  /post2:
    get:
      consumes:
      - application/json
      description: 可按社区按时间或分数排序查询帖子列表接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      - description: 可以为空,如果为空，则查询所有帖子，否则查询相应社区的帖子
        in: query
        name: community_id
        type: integer
      - description: 排序依据
        example: score
        in: query
        name: order
        type: string
      - description: 页码
        example: 1
        in: query
        name: page
        type: integer
      - description: 每页数据量
        example: 10
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponsePostList'
      security:
      - ApiKeyAuth: []
      summary: 升级版帖子列表接口
      tags:
      - 帖子相关接口
swagger: "2.0"